move regif basic access test to lib.
