+++ lib/src/test/scala/spinal/lib/bus/regif/StrbTester.scala
@@ -1,9 +1,9 @@
-package spinal.tester.generator.regif
+package spinal.lib.bus.regif
+
+import AccessType._
 
 import spinal.core._
 import spinal.core.sim._
-import spinal.lib.bus.regif._
-import spinal.lib.bus.regif.AccessType._
 import spinal.lib._
 import spinal.lib.bus.amba4.apb.sim.Apb4Driver
 import spinal.lib.bus.amba4.apb.{Apb4, Apb4Config}
